requirements.txt
python-dotenv>=1.0.1
torch>=2.3.0
transformers>=4.41.0
accelerate>=0.32.0
modelscope>=1.13.1
openai>=1.40.0
requests>=2.31.0
gradio>=4.41.0

